pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "execudeck"
version = "0.1.0"
description = "Review, generate, and edit executive PowerPoint presentations"
requires-python = ">=3.12"
dependencies = [
    "python-pptx>=0.6.23",
    "click>=8.1",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
execudeck = "execudeck.cli:cli"
